requirements.txt
flask
duckdb
pandas
schedule
google-generativeai
tabulate
python-dotenv
